Improve accessibility of tab links on groups forms
Extract the tab links from `GroupFormHeader` into a standalone component and
improve accessibility by:

 - Marking the tabs as being part of a list
 - Setting `aria-current` on the link that corresponds to the current route
